feat: add flujo de anticipos a cuentas por pagar

+++ app/Http/Controllers/AccountPayable/PayPayableController.php
@@ -3,6 +3,7 @@ namespace App\Http\Controllers\AccountPayable;
 
 use App\Http\Controllers\Controller;
 use App\Http\Resources\PayableResource;
+use App\Models\BankMovement;
 use App\Models\Payable;
 use App\Models\PayPayable;
 
@@ -17,35 +18,30 @@ class PayPayableController extends Controller
         }
 
         $payable = Payable::find($object->payable_id);
-        // if ($object->latest_bank_movement_transaction) {
-        //     if ($object->latest_bank_movement->status === "Confirmado") {
-        //         return response()->json([
-        //             'message' => 'El ingreso a caja grande ya fue confirmado y no se puede eliminar.'
-        //         ], 422);
-        //     }
-        // }
+        if ($object->latest_bank_movement_transaction) {
+            if ($object->latest_bank_movement->status === "Confirmado") {
+                return response()->json([
+                    'message' => 'El ingreso a caja grande ya fue confirmado y no se puede eliminar.'
+                ], 422);
+            }
+        }
 
-                           // $venta = Moviment::find($payable->moviment_id);
-                           // if ($venta->status == "Anulada") {
-                           //     return response()->json(['message' => 'La Venta está Anulada'], 422); // Cambiado a 404 para mejor claridad
-                           // }
         $object->delete(); // Cambia esto si se necesita otra lógica
 
         $payable->updateMontos();
-        // if ($object->latest_bank_movement_anticipo) {
+        if ($object->latest_bank_movement_anticipo) {
 
-        //     if ($object->latest_bank_movement->transaction_concept_id == 1) {
-        //         $movement_anticipo = BankMovement::find($object->latest_bank_movement->id);
-        //         $movement_anticipo->update_montos_anticipo();
-        //     }
-        // }
-
-        // if($object->latest_bank_movement_transaction){
-        //     {
-        //         $movement = BankMovement::find($object->latest_bank_movement->id);
-        //         $movement->delete();
-        //     }
-        // }
+            if ($object->latest_bank_movement->transaction_concept_id == 2) {
+                $movement_anticipo = BankMovement::find($object->latest_bank_movement->id);
+                $movement_anticipo->update_montos_anticipo_egreso();
+            }
+        }
+        if($object->latest_bank_movement_transaction){
+            {
+                $movement = BankMovement::find($object->latest_bank_movement->id);
+                $movement->delete();
+            }
+        }
         return response()->json(['payable' => new PayableResource(Payable::find($payable->id))], 200);
     }
 }
